requirements.txt
autogen==1.0.16
hpsklearn==0.1.0
joblib==1.3.2
matplotlib==3.8.2
numpy==1.23.5
pandas==2.2.2
pyautogen==0.2.29
python-dotenv==1.0.1
scikit_learn==1.5.0
